split up main into multiple files
