+++ styles.go
@@ -1,0 +1,27 @@
+package main
+
+import "github.com/charmbracelet/lipgloss"
+
+var (
+	borderStyle = lipgloss.NewStyle().
+			Border(lipgloss.NormalBorder()).
+			BorderForeground(lipgloss.Color("#874BFD")).
+			Padding(0, 1)
+
+	statusStyle = lipgloss.NewStyle().
+			Border(lipgloss.RoundedBorder()).
+			BorderForeground(lipgloss.Color("#00FF00")).
+			Padding(0, 1).
+			Width(50)
+
+	listStyle = lipgloss.NewStyle().
+			Border(lipgloss.NormalBorder()).
+			BorderForeground(lipgloss.Color("#FFD700")).
+			Margin(1, 0).
+			Padding(1)
+
+	helpStyle = lipgloss.NewStyle().
+			Foreground(lipgloss.Color("#888888")).
+			Italic(true).
+			Margin(1, 0)
+)
